// src/components/PsyPoints/types.ts
export interface AUSTaxType {
  userIsOriginalMinter: string
  userIsAUSTaxResOrigMintingPeriod?: string
  userLocAUSOrigMintingPeriod?: string
  userRegAUSGSTOrigMintingPeriod?: string
  userIsAUSTaxResident: string
  userInAUS: string
  userRegisteredAUSGST: string
  walletAddresses: string
  declareTruth: string
  referralWalletAddress?: string
}